fix:修复dish模块的一些错误

+++ model/sql/dish_flavor.go
@@ -7,7 +7,7 @@ import (
 
 type DishFlavor struct {
 	gorm.Model
-	DishID int64           `gorm:"not null;comment:'菜品'"`
+	DishID uint            `gorm:"not null;comment:'菜品'"`
 	Name   string          `gorm:"type:varchar(32);comment:'口味名称'"`
 	Value  json.RawMessage `gorm:"type:json;comment:'口味数据list'"`
 }
